requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
altair==5.2.0
joblib==1.3.2
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0 